requirements.txt
tqdm==4.40.0
pandas==1.1.1
scikit-surprise==1.1.1
click==7.1.1